Reduce image size and shift to assets
